+++ src/app/shell/rating-mode/RatingMode.tsx
@@ -7,10 +7,12 @@ import { $rootScope } from 'ngimport';
 import { D2ManifestDefinitions, getDefinitions } from '../../destiny2/d2-definitions.service';
 import { getReviewModes, D2ReviewMode } from '../../destinyTrackerApi/reviewModesFetcher';
 import { D2StoresService } from '../../inventory/d2-stores.service';
+import { getPlatformOptions } from '../../destinyTrackerApi/platformOptionsFetcher';
 
 interface State {
   open: boolean;
   reviewsModeSelection: number;
+  platformSelection: number;
   defs?: D2ManifestDefinitions;
 }
 
@@ -21,7 +23,9 @@ export default class RatingMode extends React.Component<{}, State> {
 
   constructor(props) {
     super(props);
-    this.state = { open: false, reviewsModeSelection: settings.reviewsModeSelection };
+    this.state = { open: false,
+      reviewsModeSelection: settings.reviewsModeSelection,
+      platformSelection: settings.reviewsPlatformSelection };
   }
 
   componentDidMount() {
@@ -29,7 +33,7 @@ export default class RatingMode extends React.Component<{}, State> {
   }
 
   render() {
-    const { open, reviewsModeSelection, defs } = this.state;
+    const { open, reviewsModeSelection, defs, platformSelection } = this.state;
 
     if (!defs) {
       return null;
@@ -37,18 +41,34 @@ export default class RatingMode extends React.Component<{}, State> {
 
     return (
       <div>
-          <span className="link" onClick={this.toggleDropdown} ref={this.dropdownToggler} title={t('DtrReview.ForGameMode')}>
-            <i className='fa fa fa-thumbs-up'/>
-          </span>
-          {open &&
-          <ClickOutside onClickOutside={this.closeDropdown}>
-            <div className="mode-popup">
+        <span className="link" onClick={this.toggleDropdown} ref={this.dropdownToggler} title={t('DtrReview.RatingsOptions')}>
+          <i className='fa fa fa-thumbs-up'/>
+        </span>
+        {open &&
+        <ClickOutside onClickOutside={this.closeDropdown}>
+        <div className="mode-popup">
+          <div className="mode-row">
+            <div className="mode-column">
               <label className="mode-label" htmlFor="reviewMode">{t('DtrReview.ForGameMode')}</label>
+            </div>
+            <div className="mode-column">
               <select name="reviewMode" value={reviewsModeSelection} onChange={this.modeChange}>
                 {this.reviewModeOptions.map((r) => <option key={r.mode} value={r.mode}>{r.description}</option>)}
               </select>
             </div>
-          </ClickOutside>}
+          </div>
+          <div className="mode-row">
+            <div className="mode-column">
+              <label className="mode-label" htmlFor="reviewMode">{t('DtrReview.ForPlatform')}</label>
+            </div>
+            <div className="mode-column">
+              <select name="platformSelection" value={platformSelection} onChange={this.platformChange}>
+                {getPlatformOptions().map((r) => <option key={r.description} value={r.platform}>{r.description}</option>)}
+              </select>
+            </div>
+          </div>
+        </div>
+        </ClickOutside>}
       </div>
     );
   }
@@ -79,6 +99,18 @@ export default class RatingMode extends React.Component<{}, State> {
     settings.reviewsModeSelection = newModeSelection;
     D2StoresService.refreshRatingsData();
     this.setState({ reviewsModeSelection: newModeSelection });
+    $rootScope.$broadcast('dim-refresh');
+  }
+
+  private platformChange = (e?) => {
+    if (!e || !e.target) {
+      return;
+    }
+
+    const newPlatformSelection = e.target.value;
+    settings.reviewsPlatformSelection = newPlatformSelection;
+    D2StoresService.refreshRatingsData();
+    this.setState({ platformSelection: newPlatformSelection });
     $rootScope.$broadcast('dim-refresh');
   }
 }