+++ docs/CHANGELOG.md
@@ -14,6 +14,7 @@
 * Added Bow stats "Draw Time" and "Inventory Size".
 * Disabled vendorengrams.xyz integration until they are back online.
 * Review modes - say hello to Gambit (and goodbye to Trials, at least for a little while).
+* Ratings platform selection changes made easier.
 * Added Etheric Spiral and Etheric Helix to the list of reputation items.
 
 # 4.68.3 (2018-09-03)